refactor: extract vendor search dialog
